refactor(core): Consolidate execution controllers (no-changelog)
